Moved common Azure data storage functions to download_utils.py
